package rename
 Changes to be committed:
	renamed:    fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/BlockEntityRendererRegistry.java -> fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/BlockEntityRendererRegistry.java
	renamed:    fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/EntityRendererRegistry.java -> fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/EntityRendererRegistry.java
	renamed:    fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/HudRenderCallback.java -> fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/HudRenderCallback.java
	renamed:    fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/InvalidateRenderStateCallback.java -> fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/InvalidateRenderStateCallback.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinBlockEntityRenderDispatcher.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinEntityRenderManager.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinInGameHud.java
	modified:   fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinWorldRenderer.java
