+++ fabric-rendering-v1/src/main/java/net/fabricmc/fabric/api/client/render/v1/InvalidateRenderStateCallback.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.client.render;
+package net.fabricmc.fabric.api.client.render.v1;
 
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
