+++ fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinEntityRenderManager.java
@@ -32,7 +32,7 @@ import net.minecraft.entity.Entity;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.fabricmc.fabric.api.client.render.EntityRendererRegistry;
+import net.fabricmc.fabric.api.client.render.v1.EntityRendererRegistry;
 
 @Environment(EnvType.CLIENT)
 @Mixin(EntityRenderDispatcher.class)
